requirements.txt
Django==1.10.1
Pillow==2.3.0
djangorestframework==3.4.6
mysqlclient==1.3.7
amqp==2.1.1
celery==4.0.0
requests
TwitterAPI
raven
